requirements.txt
opencv-python~=4.12.0.88
torch~=2.7.1
torchvision~=0.22.1
numpy~=2.1.3
norse~=1.1.0
tensorflow~=2.19.0
pillow~=11.3.0